Added handling for *types.Alias
